update int PL_A1

+++ src/bank.cpp
@@ -20,7 +20,7 @@ PREDICATE(createMesh, 1)
     	auto M = pd::Matrix4<float>::IDENTITY();
    	//test::registerIdealEnvIST("./../../../knowrob/knowrob_bank/src/input/framefull.log",1,M,1,"./../../../knowrob/knowrob_bank/src/input/framediff46.log");
 
-        test::registerIdealEnvIST("./../../../knowrob_bank/src/input/0.log",1,M,1,PL_A1);
+        test::registerIdealEnvIST("./../../../knowrob_bank/src/input/0.log",1,M,1,(int)PL_A1);
     
     	// Dimentional Setting
     	Vec3<float> m_min( -90, -90, -90);
